(ALC) update documentation, explicit importations,  missing config file message
Added the explicit importation of numpy and yaml (only for clarity)
Updated the docstring of the function overallSimulation (missing what the function returns if the returnRes is set to True.
Added an explicit message if the configuration file is missing
removed a couple of tailing spaces (sorry automatic spyder formatting)
Suggesting to add the script version of the jupyter notebook.